Move postEdgeData to FilterGaphService
